TAUR-1314: Moved skeleton app into tests to make it obvious it is for testing
