+++ htmengine/README.md
@@ -98,14 +98,16 @@ example:
     files = ../../htmengine/conf/supervisord-base.conf
 
 Running HTMEngine Integration Tests
------------------------
+-----------------------------------
 
 The HTM engine has a basic skeleton application exclusively for running tests.
 To run the tests, you must have MySQL, RabbitMQ, and Supervisor running.
 
-Make sure that `APPLICATION_CONFIG_PATH` is set to point to htmengine/conf
+Make sure that `APPLICATION_CONFIG_PATH` is set to point to tests/
 For example:
-`export APPLICATION_CONFIG_PATH=/Users/{name}/nta/numenta-apps/htmengine/conf
+```export APPLICATION_CONFIG_PATH=/Users/{name}/nta/numenta-apps/htmengine/tests/support/skeleton-app/conf```
+(alternatively you can use the `APPLICATION_CONFIG_PATH` from an app based off
+of the HTMEngine, just be careful to not reset an existing MySQL database)
 
 With MySQL already started start/restart RabbitMQ:
 `rabbitmq-server -detached`
@@ -118,12 +120,13 @@ rabbitmqctl start_app
 ```
 
 Setup the htmengine MySQL database:
-`htmengine-create-db`
+(don't do this if you're testing on Grok, Taurus, or another HTMEngine application)
+`./tests/integration/skeleton-app/reset_db.py`
 
 Start `supervisord`:
-`supervisord -c conf/supervisord.conf`
-*NOTE: Be sure to run `mkdir logs` within htmengine so supervisor
-has a place to store its logs*
+```supervisord -c tests/support/skeleton-app/conf/supervisord.conf```
+*NOTE: Be sure to run `mkdir logs` within the `skeleton-app` directory so
+supervisor has a place to store its logs*
 
 Run the integration tests (currently only `result_quality_test.py`):
 `py.test tests/integration/result_quality_test.py`
