+++ htmengine/tests/support/skeleton-app/conf/supervisord.conf
@@ -13,7 +13,7 @@ file=%(here)s/../htmengine-supervisor.sock   ; (the path to the socket file)
 port=127.0.0.1:9001
 
 [supervisord]
-environment=APPLICATION_CONFIG_PATH=/Users/jweiss/nta/applications/htmengine/conf
+environment=APPLICATION_CONFIG_PATH=/opt/numenta/htmengine/tests/support/conf/
 pidfile=%(here)s/../htmengine-supervisord.pid
 identifier=htmengine-supervisor
 logfile=%(here)s/../logs/htmengine-supervisord.log
